+++ imgui.ini
@@ -15,6 +15,6 @@ Collapsed=0
 
 [Window][Object Info]
 Pos=0,30
-Size=384,690
+Size=384,854
 Collapsed=0
