+++ src/gui.rs
@@ -1,6 +1,6 @@
 pub mod draw_creation_gui_sys;
 pub mod graphs;
-pub mod gui_systems;
 pub mod imgui_wrapper;
 pub mod signals;
+pub mod systems;
 pub mod ui;
